pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/StreetSolarTrack"]
[project]
name = "StreetSolarTrack"
version = "0.0.8"
authors = [
  { name="Shiqi Wang", email="qtec@outlook.com" },
]
description = "A Street View Image Processing and Solar Loaction package"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "Pillow>=10.0.1",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]



[project.urls]
Homepage = "https://github.com/GIStudio/StreetSolarTrack"
Issues = "https://github.com/GIStudio/StreetSolarTrack/issues"